+++ DCS-BIOS/EventArgs/DCSBIOSStringDataEventArgs.cs
@@ -2,7 +2,7 @@
 {
     public class DCSBIOSStringDataEventArgs : System.EventArgs                 
     {
-        public uint Address { get; init; }
+        public ushort Address { get; init; }
 
         public string StringData { get; init; }
     }
