+++ DCS-BIOS/Json/DCSBIOSControlOutput.cs
@@ -8,25 +8,25 @@ namespace DCS_BIOS.Json
         private string _type;
 
         [JsonProperty("address", Required = Required.Default)]
-        public uint Address { get; set; }
+        public ushort Address { get; set; }
 
         [JsonProperty("description", Required = Required.Default)]
         public string Description { get; set; }
 
         [JsonProperty("mask", Required = Required.Default)]
-        public uint Mask { get; set; }
+        public ushort Mask { get; set; }
 
         [JsonProperty("max_value", Required = Required.Default)]
-        public int MaxValue { get; set; }
+        public ushort MaxValue { get; set; }
 
         [JsonProperty("shift_by", Required = Required.Default)]
-        public int ShiftBy { get; set; }
+        public ushort ShiftBy { get; set; }
 
         [JsonProperty("suffix", Required = Required.Default)]
         public string Suffix { get; set; }
 
         [JsonProperty("max_length", Required = Required.Default)]
-        public int MaxLength { get; set; }
+        public ushort MaxLength { get; set; }
         
         [JsonProperty("address_identifier", Required = Required.Default)]
         public string AddressIdentifier { get; set; }
